requirements.txt
langchain
langchain-community
langchain-groq
groq
pymongo
pypdf2
sentence-transformers
python-dotenv
scikit-learn
numpy